+++ src/components/Account/AccountProfilePage.tsx
@@ -15,6 +15,7 @@ import { copyToClipboard, ethErrorHandler, formatTokens } from "lib/util";
 import * as queryString from "query-string";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
+import { Helmet } from "react-helmet";
 import { connect } from "react-redux";
 import { RouteComponentProps } from "react-router-dom";
 import { IRootState } from "reducers";
@@ -128,7 +129,6 @@ class AccountProfilePage extends React.Component<IProps, IState> {
     const [dao, accountInfo, ethBalance, genBalance] = this.props.data;
 
     const { accountAddress, accountProfile, currentAccountAddress } = this.props;
-    //const accountProfile = this.state.profile;
 
     if (!accountProfile) {
       return "Loading...";
@@ -147,6 +147,11 @@ class AccountProfilePage extends React.Component<IProps, IState> {
         <BreadcrumbsItem to={`/profile/${accountAddress}`}>
           {editing ? (accountProfile && accountProfile.name ? "Edit 3Box Profile" : "Set 3Box Profile") : "View 3Box Profile"}
         </BreadcrumbsItem>
+        <Helmet>
+          <meta name="description" content={(accountProfile.name || accountProfile.ethereumAccountAddress) + " Profile on Alchemy by DAOstack"} />
+          <meta name="og:description" content={(accountProfile.name || accountProfile.ethereumAccountAddress) + " Profile on Alchemy by DAOstack"} />
+          <meta name="twitter:description" content={(accountProfile.name || accountProfile.ethereumAccountAddress) + " Profile on Alchemy by DAOstack"} />
+        </Helmet>
 
         {this.state.showThreeBoxModal ?
           <ThreeboxModal action={this.doUpdateProfile} closeHandler={this.closeThreeboxModal} />